a2a-cart-store/requirements.txt
uagents==0.22.10
uagents-adapter==0.6.0
a2a-sdk[http-server]>=0.3.5
python-dotenv>=1.0.1
httpx>=0.27.2
python-jose[cryptography]>=3.3.0
aiohttp>=3.9.5